Refactor library folder structure
